modificacion de procedures y crud

+++ backend/app/db/procedures/sp_borrar_admin.sql
@@ -1,9 +1,21 @@
 DELIMITER $$
 CREATE PROCEDURE `sp_borrar_admin`(
-	IN p_administrador_id INT
+    IN p_administrador_id INT
 )
 BEGIN
-	DELETE FROM `administradores`
+    DECLARE admin_count INT;
+
+    SELECT COUNT(*) INTO admin_count
+    FROM `administradores`
     WHERE `administrador_id` = p_administrador_id;
+
+    IF admin_count > 0 THEN
+        DELETE FROM `administradores`
+        WHERE `administrador_id` = p_administrador_id;
+        SELECT 'El administrador fue borrado exitosamente' AS result;
+    ELSE
+        SIGNAL SQLSTATE '45000'
+        SET MESSAGE_TEXT = 'El administrador con el ID proporcionado no existe';
+    END IF;
 END$$
 DELIMITER ;
